Remove PN Activity from fdroid
