+++ app/src/main/java/org/thoughtcrime/securesms/onboarding/PNModeActivity.kt
@@ -15,11 +15,13 @@ import androidx.annotation.DrawableRes
 import dagger.hilt.android.AndroidEntryPoint
 import network.loki.messenger.R
 import network.loki.messenger.databinding.ActivityPnModeBinding
+import org.session.libsession.utilities.Device
 import org.session.libsession.utilities.TextSecurePreferences
 import org.session.libsession.utilities.ThemeUtil
 import org.thoughtcrime.securesms.ApplicationContext
 import org.thoughtcrime.securesms.BaseActionBarActivity
 import org.thoughtcrime.securesms.home.HomeActivity
+import org.thoughtcrime.securesms.home.showHomeActivity
 import org.thoughtcrime.securesms.notifications.PushManager
 import org.thoughtcrime.securesms.notifications.PushRegistry
 import org.thoughtcrime.securesms.showSessionDialog
@@ -166,14 +168,9 @@ class PNModeActivity : BaseActionBarActivity() {
             return
         }
 
-        TextSecurePreferences.setPushEnabled(this, (selectedOptionView == binding.fcmOptionView))
-        val application = ApplicationContext.getInstance(this)
-        application.startPollingIfNeeded()
+        TextSecurePreferences.setPushEnabled(this, selectedOptionView == binding.fcmOptionView)
         pushRegistry.refresh(true)
-        val intent = Intent(this, HomeActivity::class.java)
-        intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
-        intent.putExtra(HomeActivity.FROM_ONBOARDING, true)
-        show(intent)
+        showHomeActivity()
     }
     // endregion
 }
