+++ libsession/src/main/java/org/session/libsession/utilities/Device.kt
@@ -1,6 +1,7 @@
 package org.session.libsession.utilities
 
-enum class Device(val value: String, val service: String = value) {
+enum class Device(val value: String, val service: String = value, val pushAvailable: Boolean = true) {
     ANDROID("android", "firebase"),
+    FDROID("fdroid", pushAvailable = false),
     HUAWEI("huawei");
 }
